Move collapse_fixtures to utils.py.
